Add files via upload
I accidentally deleted the former sabers.html
I made several styling changes. the homepage on the "sabers link" is now able to navigate to the home page. The styling is star wars themed as well as the fonts and their respective colors. I made a few modifications to the "saber page" to show light sabers to sell. The side nav has also been removed.